using BigDecimal for dish price and order total

+++ axon-event-commons/src/main/java/com/mycompany/axoneventcommons/order/OrderItemAddedEvent.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class OrderItemAddedEvent implements OrderEvent {
     private String itemId;
     private String dishId;
     private String dishName;
-    private Float dishPrice;
+    private BigDecimal dishPrice;
     private Short quantity;
 
 }
