+++ axon-event-commons/src/main/java/com/mycompany/axoneventcommons/restaurant/RestaurantDishAddedEvent.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,6 +14,6 @@ public class RestaurantDishAddedEvent implements RestaurantEvent {
     private String restaurantId;
     private String dishId;
     private String dishName;
-    private Float dishPrice;
+    private BigDecimal dishPrice;
 
 }
